refactor: moved pothole creation away from rpcs
